+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Added
 * Support for resolving the connection strings using environment variables [#16](https://github.com/nosinovacao/name-sdk/issues/16)
+* Extension point for the decision of which ConnectionString locator to use [#33](https://github.com/nosinovacao/name-sdk/issues/33)
+* Support for resolving the connection strings using the Asp.Net Core IConfiguration in the NAME.AspNetCore package [#33](https://github.com/nosinovacao/name-sdk/issues/33)
 
 ## v1.1.1 - 2018-02-22
 ### Fixed